+++ lib/src/filter.dart
@@ -1,10 +1,7 @@
 import 'package:jetlog/src/record.dart';
 
-/// [Filter] used to filter [Record]s by criteria defined in [Filter.filter].
-abstract class Filter {
-  /// Tests whether specified [record] will be processed.
-  ///
-  /// If `false` is returned records is not permitted to be processed,
-  /// otherwise is permitted.
-  bool filter(Record record);
-}
+/// [Filter] used to filter [Record]s by criteria.
+///
+/// If `false` is returned records is not permitted to be processed,
+/// otherwise is permitted.
+typedef Filter = bool Function(Record);