+++ lib/src/filters/multi_filter.dart
@@ -3,16 +3,15 @@ import 'package:jetlog/jetlog.dart' show Filter, Record;
 /// [MultiFilter] composites multiple filters, each of enclosing filters
 /// receives incoming record. Record is discarded once one of the filters
 /// filter it out.
-class MultiFilter implements Filter {
+class MultiFilter {
   MultiFilter(this._filters);
 
   final Iterable<Filter> _filters;
 
-  @override
-  bool filter(Record record) {
+  bool call(Record record) {
     if (_filters.isNotEmpty) {
       for (final f in _filters) {
-        if (!f.filter(record)) {
+        if (!f.call(record)) {
           return false;
         }
       }