+++ lib/src/handlers/console_handler.dart
@@ -20,7 +20,7 @@ class ConsoleHandler extends Handler {
 
   @override
   void handle(Record record) {
-    if (_filter != null && !_filter.filter(record)) {
+    if (_filter != null && !_filter.call(record)) {
       return;
     }