+++ lib/src/handlers/multi_handler.dart
@@ -32,7 +32,7 @@ class MultiHandler extends Handler {
 
   @override
   void handle(Record record) {
-    if (_filter != null && !_filter.filter(record)) {
+    if (_filter != null && !_filter.call(record)) {
       return;
     }