+++ test/level_test.dart
@@ -1,4 +1,5 @@
 import 'package:test/test.dart';
+
 import 'package:jetlog/src/level.dart';
 
 void main() {