+++ test/multi_handler_test.dart
@@ -1,13 +1,12 @@
 import 'package:test/test.dart';
 
-import 'package:jetlog/jetlog.dart' show Filter, Level, Record;
+import 'package:jetlog/jetlog.dart' show Level, Record;
 import 'package:jetlog/handlers.dart' show MultiHandler, MemoryHandler;
 
 import 'package:jetlog/src/record_impl.dart' show RecordImpl;
 
-class _DebugOnlyFilter extends Filter {
-  @override
-  bool filter(Record record) => record.level == Level.debug;
+class _DebugOnlyFilter {
+  bool call(Record record) => record.level == Level.debug;
 }
 
 void main() {