+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # CHANGELOG.md
+# 0.7.0
+* `Filter` now is typedef rather than abstract class.
+  It allows plain functions to be used as filters as well as callable classes.
+
 ## v0.6.0
 * `StreamHandler` throws on `null` stream as well as `null` formatter.